Tighten up the rules regarding booleans
'y|Y' and 'n|N' as booleans! Seriously?!?!

there's really no need to be so succinct.
Also, it totally messes up structure parsing if there are
any variables named 'y', 'Y', 'n' or 'N' in the structure.
How I know? Tried creating an array of structures (an array
of coordinate points) where these points where named X, Y ,Z.
Guess which one was parsed as a boolean (and not as a natural
number as it should have been)...

Nuff' said!

Signed-off-by: Esa Koponen <esa.koponen@jotautomation.com>